# Avoid allocations in solving linear equations