also disabled gc in benchmarks.

### benchmarks on v0.4.0-dev

before:

```
benchmark: mcep
elapsed time: 10.185991948 seconds (355 MB allocated)
elapsed time: 7.223171397 seconds (656 kB allocated)
1.410182851334218 x slower than SPTK implementation
benchmark: mgcep
elapsed time: 9.287685327 seconds (89 MB allocated)
elapsed time: 6.560359736 seconds (65 kB allocated)
1.4157281407784421 x slower than SPTK implementation
```

after:
```
benchmark: mcep
elapsed time: 7.955997896 seconds (230 MB allocated)
elapsed time: 7.11089448 seconds (656 kB allocated)
1.1188462983899323 x slower than SPTK implementation
benchmark: mgcep
elapsed time: 8.378429405 seconds (79 MB allocated)
elapsed time: 6.650064232 seconds (65 kB allocated)
1.2599020629675701 x slower than SPTK implementation
```